projeto_sptrans/data_processor_business/requirements.txt
py4j==0.10.9.7
pyspark==3.5.3
python-dotenv==1.0.1